Refactor: AnswerModel 위치 변경
