+++ src/main/java/supernova/whokie/answer/service/dto/AnswerModel.java
@@ -1,4 +1,4 @@
-package supernova.whokie.answer.controller.dto;
+package supernova.whokie.answer.service.dto;
 
 import lombok.Builder;
 import supernova.whokie.user.controller.dto.UserResponse;
